refactor: Create Grigori.Api and move services to Infrastructure
Created Grigori.Api project:
- Hosts REST API endpoints (/api/search, /api/index, /api/benchmark, etc.)
- Hosts Blazor Dashboard (references Grigori.Dashboard)
- Separate from MCP functionality

Moved services to Grigori.Infrastructure:
- SearchService, IndexService, BenchmarkService
- Bm25Scorer, ReciprocalRankFusion
- Added FileSystemGlobbing package dependency

Slimmed Grigori.Mcp to MCP-only:
- Removed REST API endpoints
- Removed Dashboard references and Blazor services
- Only MCP endpoints (stdio and HTTP transport)

Updated Docker configuration:
- Dockerfile.api now builds Grigori.Api
- Dockerfile.mcp builds MCP-only Grigori.Mcp

Project structure:
- Grigori.Api: REST API + Dashboard (port 5151)
- Grigori.Mcp: MCP Server only (stdio/HTTP transport)
- Grigori.Dashboard: Shared Blazor components
- Grigori.Infrastructure: Core services (Search, Index, Benchmark)

Closes #67
