+++ src/Grigori.Api/Grigori.Api.csproj
@@ -1,16 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-
-    <!-- Assembly info -->
-    <AssemblyName>Grigori.Mcp</AssemblyName>
-    <RootNamespace>Grigori.Mcp</RootNamespace>
-    <Version>1.0.0</Version>
-    <Description>Grigori MCP Server - Semantic code search via Model Context Protocol</Description>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <RootNamespace>Grigori.Api</RootNamespace>
 
     <!-- Container optimization -->
     <PublishSingleFile>true</PublishSingleFile>
@@ -21,20 +15,16 @@
     <!-- Runtime identifiers for cross-platform builds -->
     <RuntimeIdentifiers>linux-x64;linux-arm64;win-x64;osx-x64;osx-arm64</RuntimeIdentifiers>
 
-    <!-- Invariant globalization for smaller container size (optional) -->
-    <InvariantGlobalization>false</InvariantGlobalization>
-
     <!-- Trim settings - disabled due to ONNX Runtime reflection usage -->
     <PublishTrimmed>false</PublishTrimmed>
+
+    <!-- Static web assets for Blazor -->
+    <StaticWebAssetsSelfContained>true</StaticWebAssetsSelfContained>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ModelContextProtocol" Version="0.6.0-preview.1" />
-    <PackageReference Include="ModelContextProtocol.AspNetCore" Version="0.6.0-preview.1" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\Grigori.Contracts\Grigori.Contracts.csproj" />
+    <ProjectReference Include="..\Grigori.Dashboard\Grigori.Dashboard.csproj" />
     <ProjectReference Include="..\Grigori.Database\Grigori.Database.csproj" />
     <ProjectReference Include="..\Grigori.DataAccess\Grigori.DataAccess.csproj" />
     <ProjectReference Include="..\Grigori.Infrastructure\Grigori.Infrastructure.csproj" />
